pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum-ci"
version = "0.1.0"
description = "Automated circuit analysis and comparison for quantum engineering PRs"
requires-python = ">=3.10"
dependencies = [
    "qiskit>=1.0,<2.0",
    "qiskit-aer>=0.14",
    "PyYAML>=6.0",
    "requests>=2.31",
]

[project.scripts]
quantum-ci = "quantum_ci.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["quantum_ci*"]
